Conflicts:
	project.properties
	src/com/comdo/zf_agent_a_pad/activity/Register.java
	src/com/comdo/zf_agent_a_pad/activity/TerminalApplyCreateActivity.java
	src/com/comdo/zf_agent_a_pad/trade/entity/OpeningInfos.java
	src/com/comdo/zf_agent_a_pad/util/Config.java

+++ src/com/comdo/zf_agent_a_pad/adapter/OrderAdapter.java
@@ -116,7 +116,7 @@ public class OrderAdapter extends BaseAdapter{
  		 ImageCacheUtil.IMAGE_CACHE.get(list.get(position).getOrder_goodsList().get(0).getGood_logo(), holder.im);
  		 if(OrderList.type.equals("5")){
  			holder.isshow.setVisibility(View.VISIBLE);
- 			String string=" 原价:￥"+list.get(position).getOrder_goodsList().get(0).getGood_actualprice()+" ";
+ 			String string=" 原价:￥"+(double)list.get(position).getOrder_goodsList().get(0).getGood_actualprice()+" ";
  			SpannableString sp = new SpannableString(string);
  			sp.setSpan(new StrikethroughSpan(), 0, string.length(), Spanned.SPAN_EXCLUSIVE_EXCLUSIVE);
  			holder.isshow.setText(sp);
